refactor: decompose panes.py into panes/ subpackage (9 modules)
Split the 2,348-LOC monolithic panes.py into a panes/ subpackage with
one module per settings domain: pdf, viewer, system, network, general,
processing, images, thumbnails, discovery.

The __init__.py re-exports all _build_*_pane functions, so main.py
imports remain unchanged.

Each submodule carries only the imports it actually needs.
