+++ src/studio_ui/components/settings/panes/__init__.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+from .discovery import _build_discovery_pane
+from .general import _build_general_pane
+from .images import _build_images_pane
+from .network import _build_network_pane
+from .pdf import _build_pdf_pane
+from .processing import _build_processing_pane
+from .system import _build_system_pane
+from .thumbnails import _build_thumbnails_pane
+from .viewer import _build_viewer_pane
+
+__all__ = [
+    "_build_discovery_pane",
+    "_build_general_pane",
+    "_build_images_pane",
+    "_build_network_pane",
+    "_build_pdf_pane",
+    "_build_processing_pane",
+    "_build_system_pane",
+    "_build_thumbnails_pane",
+    "_build_viewer_pane",
+]
